+++ ampieta-game-java-common/src/main/java/com/gameServer/common/resource/EquipmentPrimaryConfigResource.java
@@ -30,11 +30,11 @@ public class EquipmentPrimaryConfigResource {
     /**
      * 1级初始属性
      */
-    private int primaryGrowthInts;
+    private float primaryGrowthInts;
     /**
      * 品阶的最大等级的最大属性值
      */
-    private int primaryGrowthMaxInt;
+    private float primaryGrowthMaxInt;
     /**
      * 属性位置所属名字
      */
@@ -56,11 +56,11 @@ public class EquipmentPrimaryConfigResource {
         return growthPosName;
     }
 
-    public int getPrimaryGrowthInts() {
+    public float getPrimaryGrowthInts() {
         return primaryGrowthInts;
     }
 
-    public int getPrimaryGrowthMaxInt() {
+    public float getPrimaryGrowthMaxInt() {
         return primaryGrowthMaxInt;
     }
